Fix sonar issue

+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/KeyValueCursorBase.java
@@ -97,7 +97,7 @@ public abstract class KeyValueCursorBase extends AsyncIteratorCursor<KeyValue> i
                         context.increment(FDBStoreTimer.Counts.LOAD_SCAN_ENTRY);
                         context.increment(FDBStoreTimer.Counts.LOAD_KEY_VALUE);
                     }
-                    limitManager.reportScannedBytes(kv.getKey().length + kv.getValue().length);
+                    limitManager.reportScannedBytes((long)kv.getKey().length + (long)kv.getValue().length);
                     // Note that this mutates the pointer and NOT the array.
                     // If the value of lastKey is mutated, the Continuation class will break.
                     lastKey = kv.getKey();
